requirements.txt
pdfplumber
python-docx
nltk
sentence-transformers
chromadb